% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletANN.R
\name{WaveletAnn}
\alias{WaveletAnn}
\title{WaveletAnn}
\usage{
WaveletAnn(y, wf = "haar", J, boundary, NForecast, ...)
}
\arguments{
\item{y}{}

\item{wf}{}

\item{J}{}

\item{boundary}{}

\item{NForecast}{}

\item{...}{}
}
\value{

}
\description{
WaveletAnn
}
\examples{

}
