% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletArima.R
\name{WaveletArima}
\alias{WaveletArima}
\title{WaveletArima}
\usage{
WaveletArima(y, wf = "haar", J, boundary, NForecast, ...)
}
\arguments{
\item{y}{}

\item{wf}{}

\item{J}{}

\item{boundary}{}

\item{NForecast}{}

\item{...}{}
}
\value{

}
\description{
WaveletArima
}
\examples{

}
