% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletArima.R
\name{WaveletArima.reg}
\alias{WaveletArima.reg}
\title{WaveletArima with external regressors}
\usage{
WaveletArima.reg(y, xreg, NForecast, method = T, flag.comb = "BG", ...)
}
\arguments{
\item{y}{}

\item{xreg}{}

\item{NForecast}{}

\item{method}{}

\item{flag.comb}{}

\item{...}{}
}
\value{

}
\description{
WaveletArima with external regressors
}
\examples{

}
