% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletBMA.R
\name{WaveletBMA}
\alias{WaveletBMA}
\title{WaveletBMA}
\usage{
WaveletBMA(
  y,
  xreg,
  wf = "haar",
  J,
  boundary,
  NForecast,
  transfo = c("log", "sqrt"),
  lambda = 0.5
)
}
\arguments{
\item{y}{}

\item{xreg}{}

\item{wf}{}

\item{J}{}

\item{boundary}{}

\item{NForecast}{}
}
\value{

}
\description{
WaveletBMA
}
\examples{

}
