% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletSS.R
\name{WaveletSS}
\alias{WaveletSS}
\title{WaveletSS}
\usage{
WaveletSS(y, wf = "haar", J, boundary, NForecast, ...)
}
\arguments{
\item{y}{}

\item{wf}{}

\item{J}{}

\item{boundary}{}

\item{NForecast}{}

\item{...}{}
}
\value{

}
\description{
WaveletSS
}
\examples{

}
