% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletSS.R
\name{WaveletSS.xreg}
\alias{WaveletSS.xreg}
\title{WaveletSS with external regressors}
\usage{
WaveletSS.xreg(
  y,
  xreg,
  NForecast,
  order = 1,
  freq = 12,
  method = F,
  flag.comb = "BG",
  ...
)
}
\arguments{
\item{y}{}

\item{xreg}{}

\item{NForecast}{}

\item{order}{}

\item{freq}{}

\item{method}{}

\item{flag.comb}{}

\item{...}{}
}
\value{

}
\description{
WaveletSS with external regressors
}
\examples{

}
