% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletANN.R
\name{WaveletAnn.xreg}
\alias{WaveletAnn.xreg}
\title{WaveletAnn with external regressors}
\usage{
WaveletAnn.xreg(
  y,
  xreg,
  wf = "haar",
  J,
  boundary,
  NForecast,
  flag.comb = "BG",
  ...
)
}
\arguments{
\item{y}{}

\item{xreg}{}

\item{wf}{}

\item{J}{}

\item{boundary}{}

\item{NForecast}{}

\item{flag.comb}{}

\item{...}{}
}
\value{

}
\description{
WaveletAnn with external regressors
}
\examples{

}
